requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
faker==20.1.0
datasets==2.16.1
huggingface-hub==0.20.2
numpy==1.26.2
schedule==1.2.0
jupyter==1.0.0
psycopg2-binary==2.9.11
